requirements.txt
Django>=2.0.3
django-bootstrap4>=0.0.6
django-debug-toolbar>=1.9.1
django-filter>=1.1.0
django-mptt>=0.9.0
Markdown>=2.6.11
psycopg2-binary>=2.7.4
py-gfm>=0.1.3
